+++ ApiSuvesaPos/Datos/Models/SaldoApartado.cs
@@ -8,7 +8,7 @@ namespace Datos.Models
     public partial class SaldoApartado
     {
         public long IdApartado { get; set; }
-        public long IdCliente { get; set; }
+        public int IdCliente { get; set; }
         public double? Saldo { get; set; }
     }
 }
